hf/requirements.txt
fastapi>=0.135.0
uvicorn[standard]>=0.42.0
jinja2>=3.1.6
pydantic>=2.12.0
pydantic-settings>=2.13.0
python-multipart>=0.0.18
python-dotenv>=1.0
pyyaml>=6.0
requests>=2.33.0
typer>=0.24.0
rich>=14.0
structlog>=25.0
